requirements.txt
flask
flask_sqlalchemy
fpdf2
